Fix dtype conversion bug in object detection validation metrics
- Add .float() conversion before .mean() in torchvision model validation
- Fixes RuntimeError when torchmetrics returns integer dtype tensors
- Update panoptic-segmentation-torchvision verification note

+++ cvlization/torch/training_pipeline/object_detection/torchvision/model.py
@@ -118,7 +118,7 @@ class LitDetector(LightningModule):
 
         # Convert multi-element tensors to scalars (Lightning 2.x requires scalar values)
         scalar_metrics = {
-            k: v.mean() if v.numel() > 1 else v
+            k: v.float().mean() if v.numel() > 1 else v
             for k, v in mAP.items()
         }
